pyproject.toml
[tool.poetry]
name = "ctk_functions"
version = "0.1.0"
description = "A beautiful description."
authors = ["Reinder Vos de Wael <reinder.vosdewael@childmind.org>"]
license = "LGPL-2.1"
readme = "README.md"
packages = [{include = "ctk_functions", from = "src"}]

[tool.poetry.dependencies]
python = "~3.11"
azure-functions = "^1.19.0"
pydantic = "^2.7.1"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.1"
pydantic-settings = "^2.2.1"
cmi-docx = "^0.1.6"
azure-storage-blob = "^12.19.1"
aiofiles = "^23.2.1"
aiohttp = "^3.9.5"
pycap = "^2.6.0"
pypandoc-binary = "^1.13"
en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz"}
polars-lts-cpu = "^0.20.23"
spacy = "^3.7.4"
openai = "^1.30.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
mypy = "^1.10.0"
pre-commit = "^3.6.2"
pytest-cov = "^5.0.0"
ruff = "^0.4.4"
types-python-dateutil = "^2.9.0.20240316"
types-pytz = "^2024.1.0.20240417"
pytest-dotenv = "^0.5.2"
pytest-mock = "^3.14.0"
types-requests = "^2.31.0.20240406"
types-aiofiles = "^23.2.0.20240403"
pytest-asyncio = "^0.23.6"

[tool.poetry.group.docs.dependencies]
pdoc = "^14.4.0"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
  "tests"
]

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv"
]
line-length = 88
indent-width = 4
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = ["ANN", "D", "E", "F", "I"]
ignore = [
  "ANN101",  # self should not be annotated.
  "ANN102"  # cls should not be annotated.
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
